Move errors.h into model/.
