+++ src/ds/hashtable.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "errors.h"
+#include "model/errors.h"
 
 /*
  * Settings that control the growth behavior of a hashtable.
