+++ src/model/errors.h
