+++ tests/unit/model/object_tests.c
@@ -2,7 +2,7 @@
 #include <string.h>
 
 #include "unit.h"
-#include "errors.h"
+#include "model/errors.h"
 #include "model/object.h"
 #include "model/klass.h"
 #include "runtime/runtime.h"
